Move constants to internal package
